Simplify project, add login runner
