--- nexacro_idpw_input_js.json
@@ -1,58 +1,0 @@
-{
-  "name": "넥사크로_IDPW_입력_JS",
-  "description": "넥사크로 기반 페이지에서 ID, PW 입력 필드를 XPath로 찾고, `send_keys` 로 값을 입력한 뒤 엔터 키를 세 번 보내는 예제",
-  "exclusive": true,
-  "steps": [
-    {
-      "action": "open_url",
-      "value": "https://store.bgfretail.com/websrc/deploy/index.html"
-    },
-    {
-      "action": "wait_elements_count",
-      "by": "class_name",
-      "value": "nexainput",
-      "count": 2,
-      "timeout": 20
-    },
-    {
-      "action": "find_element",
-      "by": "xpath",
-      "value": "//input[@id='mainframe.HFrameSet00.LoginFrame.form.div_login.form.edt_id:input']",
-      "as": "id_input"
-    },
-    {
-      "action": "send_keys",
-      "target": "id_input",
-      "keys": "${LOGIN_ID}",
-      "log": "✅ ID 입력 완료"
-    },
-    {
-      "action": "find_element",
-      "by": "xpath",
-      "value": "//*[@id='mainframe.HFrameSet00.LoginFrame.form.div_login.form.edt_pw:input']",
-      "as": "pw_input"
-    },
-    {
-      "action": "send_keys",
-      "target": "pw_input",
-      "keys": "${LOGIN_PW}",
-      "log": "✅ 비밀번호 입력 완료"
-    },
-    {
-      "action": "send_keys",
-      "target": "pw_input",
-      "keys": [
-        {"key": "ENTER", "delay": 1},
-        {"key": "ENTER", "delay": 1},
-        {"key": "ENTER"}
-      ],
-      "log": "✅ 비밀번호 입력 후 엔터 3회 입력 (1초 간격)"
-    },
-    {
-      "action": "sleep",
-      "seconds": 2
-    }
-  ],
-  "id_xpath": "//*[@id=\"mainframe.HFrameSet00.LoginFrame.form.div_login.form.edt_id:input\"]",
-  "pw_xpath": "//*[@id=\"mainframe.HFrameSet00.LoginFrame.form.div_login.form.edt_pw:input\"]"
-}
