--- structure/login_structure.json
@@ -1,6 +1,0 @@
-{
-  "url": "https://store.bgfretail.com/websrc/deploy/index.html",
-  "id_selector": "input[type='text']",
-  "password_selector": "input[type='password']",
-  "submit_selector": "button[type='submit']"
-}
